HevSock: Break busy loop and log error if requested address is unavailable.
On MSYS2, when the requested address is unavailable, connect() may repeatedly
return EINPROGRESS instead of the expected EADDRNOTAVAIL. This can cause the
program to enter a busy loop without reporting the underlying error.
As a workaround, we impose a retry limit of 100 attempts --
if exceeded, an EADDRNOTAVAIL error is raised to break the loop and signal
the failure.